Split tests into a separate module
